Move Image.kt and Layout.kt to cmera-core so that we can reuse the extensions and cropping utils
