+++ stripecardscan/src/androidTest/java/com/stripe/android/stripecardscan/framework/image/BitmapExtensionsTest.kt
@@ -5,6 +5,8 @@ import android.util.Size
 import androidx.core.graphics.drawable.toBitmap
 import androidx.test.filters.SmallTest
 import androidx.test.platform.app.InstrumentationRegistry
+import com.stripe.android.camera.framework.image.crop
+import com.stripe.android.camera.framework.image.scale
 import com.stripe.android.stripecardscan.test.R
 import org.junit.Test
 import kotlin.test.assertEquals
