+++ stripecardscan/src/androidTest/java/com/stripe/android/stripecardscan/payment/ml/SSDOcrTest.kt
@@ -4,10 +4,10 @@ import android.util.Size
 import androidx.core.graphics.drawable.toBitmap
 import androidx.test.filters.MediumTest
 import androidx.test.platform.app.InstrumentationRegistry
+import com.stripe.android.camera.framework.image.scale
+import com.stripe.android.camera.framework.image.size
+import com.stripe.android.camera.framework.util.toRect
 import com.stripe.android.stripecardscan.framework.ResourceFetcher
-import com.stripe.android.stripecardscan.framework.image.scale
-import com.stripe.android.stripecardscan.framework.image.size
-import com.stripe.android.stripecardscan.framework.util.toRect
 import com.stripe.android.stripecardscan.test.R
 import kotlinx.coroutines.runBlocking
 import org.junit.Test
