+++ stripecardscan/src/main/java/com/stripe/android/stripecardscan/framework/api/StripeApi.kt
@@ -4,6 +4,12 @@ package com.stripe.android.stripecardscan.framework.api
 import android.util.Log
 import android.util.Size
 import androidx.annotation.CheckResult
+import com.stripe.android.camera.framework.image.constrainToSize
+import com.stripe.android.camera.framework.image.crop
+import com.stripe.android.camera.framework.image.determineViewFinderCrop
+import com.stripe.android.camera.framework.image.size
+import com.stripe.android.camera.framework.image.toJpeg
+import com.stripe.android.camera.framework.util.move
 import com.stripe.android.stripecardscan.cardimageverification.SavedFrame
 import com.stripe.android.stripecardscan.framework.Config
 import com.stripe.android.stripecardscan.framework.NetworkConfig
@@ -20,16 +26,10 @@ import com.stripe.android.stripecardscan.framework.api.dto.VerificationFrameData
 import com.stripe.android.stripecardscan.framework.api.dto.VerifyFramesRequest
 import com.stripe.android.stripecardscan.framework.api.dto.VerifyFramesResult
 import com.stripe.android.stripecardscan.framework.api.dto.ViewFinderMargins
-import com.stripe.android.stripecardscan.framework.image.constrainToSize
-import com.stripe.android.stripecardscan.framework.image.crop
-import com.stripe.android.stripecardscan.framework.image.size
-import com.stripe.android.stripecardscan.framework.image.toJpeg
 import com.stripe.android.stripecardscan.framework.util.AppDetails
 import com.stripe.android.stripecardscan.framework.util.Device
 import com.stripe.android.stripecardscan.framework.util.b64Encode
-import com.stripe.android.stripecardscan.framework.util.move
-import com.stripe.android.stripecardscan.framework.util.scaleAndCenterWithin
-import com.stripe.android.stripecardscan.payment.determineViewFinderCrop
+import com.stripe.android.camera.framework.util.scaleAndCenterWithin
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
