+++ stripecardscan/src/main/java/com/stripe/android/stripecardscan/framework/image/ImageExtensions.kt
@@ -9,6 +9,7 @@ import android.renderscript.RenderScript
 import androidx.annotation.CheckResult
 import com.stripe.android.camera.framework.exception.ImageTypeNotSupportedException
 import com.stripe.android.camera.framework.image.NV21Image
+import com.stripe.android.camera.framework.image.crop
 
 /**
  * Determine if this application supports an image format.
