+++ stripecardscan/src/main/java/com/stripe/android/stripecardscan/payment/ml/CardDetect.kt
@@ -4,15 +4,15 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.Rect
 import android.util.Size
+import com.stripe.android.camera.framework.image.cropCameraPreviewToSquare
+import com.stripe.android.camera.framework.image.hasOpenGl31
+import com.stripe.android.camera.framework.image.scale
 import com.stripe.android.stripecardscan.framework.FetchedData
 import com.stripe.android.stripecardscan.framework.image.MLImage
-import com.stripe.android.stripecardscan.framework.image.scale
 import com.stripe.android.stripecardscan.framework.image.toMLImage
 import com.stripe.android.stripecardscan.framework.ml.TFLAnalyzerFactory
 import com.stripe.android.stripecardscan.framework.ml.TensorFlowLiteAnalyzer
 import com.stripe.android.stripecardscan.framework.util.indexOfMax
-import com.stripe.android.stripecardscan.payment.cropCameraPreviewToSquare
-import com.stripe.android.stripecardscan.payment.hasOpenGl31
 import org.tensorflow.lite.Interpreter
 import java.nio.ByteBuffer
 import kotlin.math.max
